# docs/swagger.yaml
basePath: /
definitions:
  achievements.Achievement:
    properties:
      created_at:
        type: string
      earned_at:
        type: string
      external_id:
        type: string
      id:
        type: string
      kind:
        $ref: '#/definitions/achievements.Kind'
      name:
        type: string
      progress_pct:
        type: number
      updated_at:
        type: string
    type: object
  achievements.Kind:
    enum:
    - badge
    - challenge
    type: string
    x-enum-varnames:
    - KindBadge
    - KindChallenge
  achievements.upsertRequest:
    properties:
      earned_at:
        type: string
      external_id:
        type: string
      kind:
        type: string
      name:
        type: string
      progress_pct:
        type: number
    type: object
...
  devices.Device:
    properties:
      battery_pct:
        type: number
      created_at:
        type: string
      display_name:
        type: string
      external_id:
        type: string
      firmware_version:
        type: string
      id:
        type: string
      last_sync_at:
        type: string
      model:
        type: string
      updated_at:
        type: string
    type: object
  devices.upsertRequest:
    properties:
      battery_pct:
        type: number
      display_name:
        type: string
      external_id:
        type: string
      firmware_version:
        type: string
      last_sync_at:
        type: string
      model:
        type: string
    type: object
...
info:
  contact: {}
  description: Personal nutrition tracking REST API. All endpoints under /products,
    /meals, and /summary require a bearer token.
  title: Nutrition API
  version: 0.1.0
paths:
  /achievements:
    get:
      parameters:
      - description: 'Filter by kind: badge | challenge'
        in: query
        name: kind
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{ achievements: [...] }'
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List achievements
      tags:
      - achievements
    post:
      consumes:
      - application/json
      description: Creates or updates an earned badge or ad-hoc challenge, upserting
        by `external_id`. Coaching context only; never feeds nutrition computation.
        Personal records live in the separate personal-records capability.
      parameters:
      - description: Optional client-supplied idempotency key
        in: header
        name: Idempotency-Key
        type: string
      - description: Achievement
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/achievements.upsertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: UPDATE (external_id already present)
          schema:
            $ref: '#/definitions/achievements.Achievement'
        "201":
          description: INSERT
          schema:
            $ref: '#/definitions/achievements.Achievement'
        "400":
          description: external_id_required | kind_invalid | name_required | progress_pct_invalid
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upsert an achievement (by Garmin badge/challenge id)
      tags:
      - achievements
...
  /devices:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: '{ devices: [...] }'
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List device records
      tags:
      - devices
    post:
      consumes:
      - application/json
      description: Creates or updates a paired Garmin device, upserting by `external_id`.
        Slowly-changing inventory — a fresh sync advances last_sync_at/battery. Reference
        context only; never feeds nutrition computation.
      parameters:
      - description: Optional client-supplied idempotency key
        in: header
        name: Idempotency-Key
        type: string
      - description: Device record
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/devices.upsertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: UPDATE (external_id already present)
          schema:
            $ref: '#/definitions/devices.Device'
        "201":
          description: INSERT
          schema:
            $ref: '#/definitions/devices.Device'
        "400":
          description: external_id_required | display_name_required | battery_pct_invalid
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upsert a device record (by Garmin device id)
      tags:
      - devices
  /devices/{id}:
    get:
      parameters:
      - description: Device UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/devices.Device'
        "404":
          description: device_not_found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a device record by id
      tags:
      - devices
...
securityDefinitions:
  BearerAuth:
    description: Static bearer token. Use the format `Bearer <token>` with either
      MOBILE_API_TOKEN or AGENT_API_TOKEN.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"